# gcs: Pool MatchAny data allocations